Add some minor tweaks required for our use-case
reverse-clients can include a "payload" that is passed to the server as
a HTTP header. The reverse-server emits an event called 'ready' when a
client connects, providing the reverse tunnel address and the client
"payload" as arguments.

We want to set up clients to connect automatically, and the server to
allocate a random port for each client (this was already possible by
setting port to 0). We need the server so emit information about each
client to the application, so that we can keep a log of which client is
available on which port.