Create TCP connection before reverse WS connection
I was having trouble with data packet getting lost immediately after
making a connection through the reverse tunnel (when using the tunnel
for SSH, this resulted in "protocol mismatch" errors). The initial
packet on the WS connection was not received by the reverse client
(the SSH server).

By reversing the order in which the sockets are connected, I eliminated
the problem. Maybe the pause mechanism was not working for the WS
connection, I'm not sure. Anyhow, by setting up the TCP connection
first, we get into the exact same state that we have in the
reverse-server, where the TCP socket is paused while waiting for the WS
connection to become ready.